docs: remove pull secret instructions

+++ docs/quickstart.sh
@@ -1,47 +1,50 @@
 #!/usr/bin/env bash
-read -p "Upbound username: " up_user; read -sp "Upbound password: " up_pass; echo ""; read -p "AWS access_key_id: " aws_access_key; read -sp "AWS secret_access_key: " aws_secret_key; export AWS_KEY=$aws_access_key; export AWS_SECRET=$aws_secret_key; export UP_USER=$up_user; export UP_PASS=$up_pass; echo "";
+set -eE
 
-VAR=$(head -n 4096 /dev/urandom | openssl sha1 | tail -c 14)
+read -p "AWS access_key_id: " aws_access_key; read -sp "AWS secret_access_key: " aws_secret_key; export AWS_KEY=$aws_access_key; export AWS_SECRET=$aws_secret_key;
 
-curl -sL "https://cli.upbound.io" | sh
+if ! kubectl -n upbound-system get deployment crossplane > /dev/null 2>&1
+then
+  printf "\n\nInstalling up CLI...\n"
+  curl -sL "https://cli.upbound.io" | sh
+  sudo mv up /usr/local/bin/
+  printf "\n\nInstalling UXP...\n"
+  up uxp install
+fi
 
-sudo mv up /usr/local/bin/
-
-up uxp install
-
-printf "\n\nChecking UXP install (this only takes a minute)..." ; while true ; do if [[ $(kubectl get deployment -n upbound-system -o jsonpath='{.items[*].status.conditions[*].status}') = "True True True True True True True True" ]]; then printf "\nUXP is ready.\n\n" ; break; else printf  "."; sleep 2; fi done
-
-up login -u $UP_USER -p $UP_PASS
+printf "\n\nChecking the UXP installation (this only takes a minute)...\n"
+kubectl -n upbound-system wait deployment crossplane --for=condition=Available --timeout=180s
 
 
-if [[ $(up org list | wc -l) = 2 ]]; then ORG=$(up org list | awk '{ print $1 }' | sed '1d') ; else up org create my-org-$VAR ; ORG=my-org-$VAR; fi
-
-up robot create my-robot-$VAR -a $ORG
-
-up robot token create my-robot-$VAR my-token-$VAR --output=token.json -a $ORG
-
-up controlplane pull-secret create package-pull-secret -f token.json
-
+printf "\n\nInstalling the provider (this will take a few minutes)...\n"
 cat <<EOF | kubectl apply -f -
 apiVersion: pkg.crossplane.io/v1
 kind: Provider
 metadata:
   name: provider-aws
 spec:
-  package: xpkg.upbound.io/upbound/provider-aws:v0.10.0
-  packagePullSecrets:
-    - name: package-pull-secret
+  package: xpkg.upbound.io/upbound/provider-aws:v0.17.0
 EOF
+kubectl wait "providers.pkg.crossplane.io/provider-aws" --for=condition=Installed --timeout=180s
+kubectl wait "providers.pkg.crossplane.io/provider-aws" --for=condition=Healthy --timeout=180s
 
-printf "\n\nChecking provider (this will take a few minutes)..." ; while true ; do if [[ $(kubectl get provider -o jsonpath='{.items[*].status.conditions[*].status}') = "True True" ]]; then printf "\n\n The provider is ready.\n\n" ; break; else echo  -n "."; sleep 3; fi done
-
-cat <<EOF > aws.txt
+creds=$(
+cat <<EOF | base64
 [default]
 aws_access_key_id = $AWS_KEY
 aws_secret_access_key = $AWS_SECRET
 EOF
+)
 
-kubectl create secret generic aws-secret -n upbound-system --from-file=creds=./aws.txt
+cat <<EOF | kubectl apply -f -
+apiVersion: v1
+kind: Secret
+metadata:
+  name: aws-secret
+  namespace: upbound-system
+data:
+  creds: ${creds}
+EOF
 
 cat <<EOF | kubectl apply -f -
 apiVersion: aws.upbound.io/v1beta1
@@ -57,18 +60,17 @@ spec:
       key: creds
 EOF
 
-cat <<EOF | kubectl apply -f -
+cat <<EOF | kubectl create -f -
 apiVersion: s3.aws.upbound.io/v1beta1
 kind: Bucket
 metadata:
-  name: upbound-bucket-$VAR
+  generateName: upbound-bucket-
 spec:
   forProvider:
     region: us-east-1
-  providerConfigRef:
-    name: default
 EOF
 
-printf "\n\nChecking AWS bucket creation (this only takes a minute)..." ; while true ; do if [[ $(kubectl get buckets -o jsonpath='{.items[*].status.conditions[*].status}') = "True True" ]]; then printf "\nYour bucket is ready.\n\n" ; break; else printf  "."; sleep 2; fi done
+printf "\n\nChecking AWS bucket creation (this only takes a minute)...\n"
+kubectl wait "$(kubectl get buckets -o name)" --for=condition=Ready --timeout=180s
 
 kubectl get buckets
